Better Physics 2D Driver, Clang support added.
2D: TileMap now create bodies per physics tile.
Core: Renaming QueueArray to FreeList, fixing memory corruption in HashMap::resize and StringMap::resize, new static collection StaticFreeList.
OS: Semaphore now use a OS native implementation.
Physics: Reworkd for Area2D and Body2D, prefer physics object with only one shape instead of multiple ones, first implementation of a tile based collisions check.

+++ src/physics/physics_2d.cpp
@@ -20,7 +20,7 @@ void Physics2D::initialize(const mem::Allocator& allocator, Physics2D::DriverTyp
     data.properties = StringMap<PropertyValue>::with_size(data.allocator, 4);
     data.properties.insert("/gravity", Property::Vector2(0, -98));
     data.properties.insert("/debug_draw", Property::Bool(false));
-    data.properties.insert("/tile_size", Property::Integer(100));
+    data.properties.insert("/tile_size", Property::Integer(64));
 
     vtable.initialize(allocator);
 }
